Add Webpack to the project.
The main script file and the level editor file were so tangled
that I couldn't figure out what to export from the latter, so I
simply copied it's contents to the main script.

The Webpack entry point is the default src/index.js and the
bundle file is also the default, so no explicit Webpack config
file is needed.

The bundled script is included in the repo to make it playable
on the GH pages.